Fix circular dependency in checkWorkspaceActivity
